カスパールに効果のある武器を設定できるようにしました。

+++ MagicBullet/Assets/FUJIYOSHI/Scripts/Interfaces/IEnemy.cs
@@ -7,7 +7,7 @@ public interface IEnemy
 {
     // ダメージを受ける
     // 1:使用武器の番号2:ダメージ量
-    public void EnemyDamage(float Damage);
+    public void EnemyDamage(int Damage);
 
     // 攻撃値返却
     // 戻り値:ダメージ量
